@@ -21,12 +21,13 @@
         "debril/feed-io": "~3.0|~4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.5",
+        "phpunit/phpunit": "~7.0",
         "doctrine/common": "~2.3",
         "doctrine/doctrine-bundle": "~1.2",
         "symfony/validator": ">3.0",
         "symfony/finder": ">3.0",
-        "symfony/browser-kit": ">3.0"
+        "symfony/browser-kit": ">3.0",
+        "symfony/yaml": "^4.0"
     },
     "suggest": {
         "doctrine/doctrine-bundle": "allow to provide feeds' content from database"
