@@ -2,7 +2,9 @@
 
 namespace Debril\RssAtomBundle\DependencyInjection;
 
-class ConfigurationTest extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class ConfigurationTest extends TestCase
 {
 
     public function testGetConfigTreeBuilder()
