@@ -1,11 +1,11 @@
 <?php
 
 namespace Debril\RssAtomBundle\Provider;
-
+use PHPUnit\Framework\TestCase;
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.0 on 2013-01-25 at 23:47:50.
  */
-class MockProviderTest extends \PHPUnit_Framework_TestCase
+class MockProviderTest extends TestCase
 {
     /**
      * @var MockProvider
